requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
requests==2.32.3
greenlet==1.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
nose==1.3.7
pydantic==1.10.2
typing_extensions==4.3.0
Werkzeug==2.2.2
zipp==3.8.1
python-dotenv==1.0.1
requests-toolbelt==1.0.0
boxsdk==1.3.0